requirements.txt
flask
python-dotenv
openai
requests
PyGithub
PyJWT 
cachetools